rewriter waits for module info and attaches dep graph
Fixed bug: the rewriter was treating the result of ModuleSet.fetch
as a ResolvedModule when it is UnresolvedModule|Future.
Reworked to await resolution.

Added dep and rdep fields to interface Module and subtypes.

Modified rewriter to convert edges from type Job instances
into dep and rdep lists in the RewrittenModules it creates.